uploadFile: clear File value after upload complete
Set the 'value' of the file upload DOM element to 'null' after the file
is complete, in order to allow a file to be uploaded repeatedly.

We use a <file> element as a backing store to allow users to upload
files.  Without clearning this value, it is not possible to upload a
file repeatedly without changing the filename.

This fixes issue #12.

Signed-off-by: Sean Cross <sean@xobs.io>