allow re-ordering of examples based on directory name
strict alphabetical ordering of examples can lead to confusion
among neophytes. For example, for the Chibi Scope, "Animation"
is the first thing, not "Hello World", according to alphabetical
sorting, which can be confusing.

Modify the auto-build code so that order is based on directory name.
Since directory name is decoupled from example name, one can
defensively name directories, e.g. if examples are developed out
of order you can give them an increment that is non-monotonic
and users won't see this.