logo: have "chibi-tronics" vanish if window gets small
If the window width drops below 480px, hide the text "chibi-tronics".

We do this by splitting the logo into two halves and adding a CSS rule
to hide the second half if the window shrinks.

This patch touches the CSS file (to add the rule), the base HTML (to
replace the image with two halves), and the images directory (to provide
the new images).

Additionally, we now pre-size the images, saving about 3kB.

Finally, because we're manipulating the button box area, we also fix a
separate bug where the buttons don't have the correct padding.

Signed-off-by: Sean Cross <sean@xobs.io>