html: test: add page to test audio support
Add a simple page to test a browser's compatibility.

This copies a lot of code from ltc-npm-modulate.  We might be able to
get away with including the main index.js file, however minification,
space, and time constraints make that prohibative.

We do reuse the css file, however.

This page will detect if Javascript is disabled, unsupported Javascript,
and incomplete playback.

Signed-off-by: Sean Cross <sean@xobs.io>